requirements.txt
signalwire-agents
requests
python-dotenv
